Change to the resource structure
The resource hidden directory used to contain three files: one for the
gateways, one for the masters and one for the nodes.
Now, this directory contains two keywords resource files relative to the
connection to the cluster and the state of the kubernetes cluster. The
last file in the directory is hidden and contains the infrastrcuture
adresses.